+++ include/base/container/ISet.h
@@ -42,9 +42,9 @@ namespace base
 		///
 		/// @brief 元素个数。
 		///
-		/// @return int32_t
+		/// @return int64_t
 		///
-		virtual int32_t Count() const = 0;
+		virtual int64_t Count() const = 0;
 
 		using base::IEnumerable<ItemType const>::GetEnumerator;
